pyproject.toml
[project]
name = "influflow"
version = "0.0.15"
description = "influflow"
authors = [
    { name = "influflow" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.4.8,<0.5.0",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.3.22,<0.4.0",
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.6",
    "langchain-deepseek>=0.1.2",
    "langchain-tavily",
    "langchain-groq>=0.2.4",
    "openai>=1.86.0,<1.90.0",
    "tavily-python>=0.5.0",
    "arxiv>=2.1.3",
    "pymupdf>=1.25.3",
    "xmltodict>=0.14.2",
    "linkup-sdk>=0.2.3",
    "duckduckgo-search>=3.0.0",
    "exa-py>=1.8.8",
    "requests>=2.32.3",
    "beautifulsoup4==4.13.3",
    "python-dotenv>=1.0.1",
    "pytest",
    "httpx>=0.24.0",
    "markdownify>=0.11.6",
    "azure-identity>=1.21.0",
    "azure-search>=1.0.0b2",
    "azure-search-documents>=11.5.2",
    "rich>=13.0.0",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.3.45,<0.4.0",
    "streamlit>=1.46.0,<1.50.0",
    # FastAPI相关依赖
    "fastapi>=0.104.0,<0.120.0",
    "uvicorn[standard]>=0.24.0,<0.30.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["influflow"]

[tool.setuptools.package-dir]
"influflow" = "src/influflow"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
